Add dedicated module for `Vertices`
