+++ src/kernel/shape/mod.rs
@@ -1,8 +1,12 @@
+pub mod vertices;
+
 use kiddo::KdTree;
 
 use crate::math::{Point, Scalar};
 
-use super::topology::{edges::Edges, faces::Faces, vertices::Vertex};
+use super::topology::{edges::Edges, faces::Faces};
+
+use self::vertices::Vertices;
 
 /// The boundary representation of a shape
 ///
@@ -34,30 +38,6 @@ impl Shape {
         Vertices {
             vertices: &mut self.vertices,
         }
-    }
-}
-
-/// The vertices of a shape
-pub struct Vertices<'r> {
-    vertices: &'r mut VerticesInner,
-}
-
-impl Vertices<'_> {
-    /// Create a vertex
-    ///
-    /// The caller must make sure to uphold all rules regarding vertex
-    /// uniqueness.
-    ///
-    /// # Implementation note
-    ///
-    /// This method is the only means to create `Vertex` instances, outside of
-    /// unit tests. That puts this method is in a great position to enforce
-    /// vertex uniqueness rules, instead of requiring the user to uphold those.
-    pub fn create(&mut self, point: Point<3>) -> Vertex {
-        self.vertices
-            .add(&point.into(), point)
-            .expect("Error adding vertex");
-        Vertex(point)
     }
 }
